Fixes #333 

Benchmark:
  - updated to reflect essential use cases and performance eaters
  - NOTE: benchmark framework reports incorrect high number for first bench on some runs, reason unknown

Based on my own usage of chalk, I select KPM:
  - `cached`: average of `cached: 1 style` and `cached: 2 styles` metrics
  - `non-cached`: average of `1 style` and `2 styles` metrics
  - `total`: average of all metrics

KPM increase:
  - `cached`: x23
  - `non-cached`: x129
  - `total`: x24

Changes(incrementally increasing preformance):
  - replaced regexp replaces by indexOf-slice implementation
  - streamlined builder instantiation
  - moved argument handling out of hot zone and optimized
  - removed all regexp stuff and escape-string-regexp dependency
  - replaced arrays by linked lists
  - precomputed chains of open/close codes
  - close code replace not applied to codes inserted around CRLF, it can change output in some cases, but doesnt change effective styles
  - added fastpath for do no close code replaces if there is no escape sequences in source string
  - cached simple getters results(in getters `this`)

Original(with new benchmark) metrics:
```
         269,472 op/s » 1 style
         130,959 op/s » 2 styles
          86,590 op/s » 3 styles
       1,407,114 op/s » cached: 1 style
         749,141 op/s » cached: 2 styles
         502,693 op/s » cached: 3 styles
       1,245,477 op/s » cached: 1 style with newline
       1,386,202 op/s » cached: 1 style nested intersecting
       1,368,144 op/s » cached: 1 style nested non-intersecting
```
Original KPM:
  - `cached`: 1078 kop/s
  - `non-cached`: 200 kop/s
  - `total`: 794 kop/s

Result metrics:
```
      29,685,124 op/s » 1 style
      21,955,435 op/s » 2 styles
      18,687,591 op/s » 3 styles
      26,847,185 op/s » cached: 1 style
      23,004,420 op/s » cached: 2 styles
      18,772,011 op/s » cached: 3 styles
      12,552,691 op/s » cached: 1 style with newline
       5,585,706 op/s » cached: 1 style nested intersecting
      15,220,157 op/s » cached: 1 style nested non-intersecting
```
Result KPM:
  - `cached:` 24925 kop/s
  - `non-cached`: 25820 kop/s
  - `total`: 19145 kop/s