@stroncium 2019-03-27
`cached: 1 style nested intersecting` might be a way to improve the speed further. It hits replaces of `close` codes with `open` codes, which trashes strings a lot.