@stroncium 2019-03-27
Another insight is that for full benchmark, time is spent:
 - 6.9% + sum of some small numbers = probably around 15% on benchmark own code
 - 8.3% + 6.5% = 14.8% on operations related to setting prototypes
 - 10.4% on unoptimized indexOf(which currently happens when replacing close codes)
 - 7.2% + 2.7% = 9.9% on traversing prototypes
 - 5.4% on optimized indexOf, 6.5% on string concatenation

Which shows us that:
 - using better benchmark framework might help with obtaining clearer results
 - caching might not work entirely as planned(which is what I want to check next)
 - replacing close codes might need more work, small state machine for that could probably speed stuff up a lot, hugely increasing performance in `1 style nested non-intersecting` and especially in `1 style nested intersecting`
 - complex prototypes might deoptimize cache retrieval from objects, so it might be effective to precompute low nesting level builders and deploy them without using complex prototype, also proxy stuff might be solution too(but I didn't test it yet)
 - at least 12% of time is spent on operations we MUST perform (which is in line with my reference tests which show that performing checks and generating resulting strings for simple 1 style application is roughly 8 times faster than what is done by chalk), which is quite good result, though some of this time might be spent on repetitive operations.