@stroncium 2019-03-27
Also, providing shorthands like `redBold`(and using them) is supposed to increase performance by ~20%