@sindresorhus 2019-05-12
This is really awesome! 🙌

Are there any more optimization opportunities you can think of? Or are you done?