@sindresorhus 2019-05-12
Can you give the PR a proper descriptive title?