@sindresorhus 2019-07-12
Ok. Hehe. No worries.