@aminya 2021-01-08
@stroncium Looking at this a year later, something took my attention. 😄  

You have written `replaced arrays by linked lists` as a performance improvement change. FYI, a linked list is horrible in terms of performance when compared to dense data structures. I could not find the change you are referring to, but I just wanted to mention this. Sorry, if I am too late to the party.