@Qix- 2021-01-08
That's a blanket statement and not very constructive. Linked lists are useful and can very much be performing depending on how they're used. Otherwise, nobody would use them.

If you have a criticism of a specific piece of code, please go into detail. Pointing someone to an absolutist video about data structures when clearly the OP knows a bit of what they're doing can be seen as condescending. Further, it doesn't provoke conversation.